requirements.txt
streamlit
pandas
pycaret
streamlit_lottie
Pillow
Image
requests
sklearn.ensemble